requirements.txt
-e svn+http://code.djangoproject.com/svn/django/branches/releases/1.4.X#egg=django
-e git+git://github.com/wichitacode/django_utils.git@v1.0#egg=django_utils

django-ses
pydkim
dnspython
